pyproject.toml
[tool.poetry]
name = "arataso"
version = "0.1.0"
description = ""
authors = ["新篠ゆう <37145593+yuarasino@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
websocket_client = "^0.57.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-mock = "^3.5.1"
pytest-randomly = "^3.5.0"
flake8 = "^3.8.4"
black = "^20.8b1"
flake8-black = "^0.2.1"
isort = "^5.7.0"
flake8-isort = "^4.0.0"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_level = "INFO"
testpaths = ["tests"]

[tool.black]
line-length = 99
exclude = '(\.git|\.devcontainer|\.venv|\.pytest_cache|\.mypy_cache)'

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3
skip = '.git,.devcontainer,.venv,.pytest_cache,.mypy_cache'
